Initial code set-up for registration
Set up the initial base for the registration class. Created some basic
variables and user input for the information needed to successfully
register. Need to discuss with team on how the information will be
relayed back to be used to create a new customer/user object (still a
bit rusty on java). Validation etc will be done after that is sorted